refactor: complete Part 3 Phase 2 — remaining shared adapters (#1457, #1227, #1242, #1371)
Extract three categories of duplicated code into shared modules:

1. BridgeContextState (#1227): Per-context ShadowRegistry + SenderKeyStore
   state was defined identically in PyO3, NAPI, and UniFFI bridges. Extracted
   to scp-ffi-common::bridge_state with shared registry and cleanup functions.

2. Attestation construction (#1457): Identity link attestation building logic
   (parse method, compute ID, build struct, validate structure, compute
   canonical bytes) was duplicated across all three bridges. Extracted to
   scp-ffi-common::attestation::build_unsigned_attestation, leaving only
   bridge-specific custody access and signing in each bridge.

3. Relay/node startup (#1242): env_or, relay_config_from_env, storage_from_env,
   init_tracing, health_check, and shutdown_signal were copy-pasted between
   scp-relay and scp-node binaries. Extracted to scp-transport::startup behind
   a new `startup` feature flag. Both binaries now delegate to the shared
   module, removing ~570 lines of duplication.

Note: #1371 (PyO3 validate.rs migration) was already completed — PyO3
validate.rs delegates to scp-ffi-common::validate since a prior PR.
