+++ crates/scp-ffi/common/Cargo.toml
@@ -29,6 +29,7 @@ scp-identity = { path = "../../scp-identity", optional = true }
 scp-transport = { path = "../../scp-transport", features = ["redb-blob"], optional = true }
 scp-node = { path = "../../scp-node", features = ["allow_unencrypted_storage"], optional = true }
 scp-platform = { path = "../../scp-platform", features = ["testing", "filesystem", "file"], optional = true }
+dashmap = { workspace = true }
 zeroize = { workspace = true }
 serde_json = { workspace = true }
 sha2 = { workspace = true }
