+++ crates/scp-ffi/napi/src/bridge_connector.rs
@@ -8,10 +8,9 @@
 //!
 //! See spec section 12 (Bridge System) and ADR-023.
 
+use scp_ffi_common::bridge_state::{BridgeContextState, bridge_state_registry};
 use scp_ffi_common::error_codes as codes;
-use std::sync::OnceLock;
 
-use dashmap::DashMap;
 use napi_derive::napi;
 
 use scp_core::bridge::provenance::{evaluate_trust_level, mark_bridge_provenance};
@@ -31,35 +30,9 @@ use crate::error::ScpNapiError;
 // ---------------------------------------------------------------------------
 // Per-context bridge state — persistent ShadowRegistry + SenderKeyStore
 // ---------------------------------------------------------------------------
-
-/// Per-context bridge connector state that persists across function calls.
-///
-/// Without this, `bridge_create_shadow` would create ephemeral
-/// `ShadowRegistry` and `SenderKeyStore` instances that are dropped when the
-/// function returns, losing all shadow identity and sender key state.
-///
-/// Keyed by context ID in `BRIDGE_STATE`.
-struct BridgeContextState {
-    shadow_registry: ShadowRegistry,
-    sender_key_store: SenderKeyStore,
-}
-
-/// Process-global registry of per-context bridge connector state.
-///
-/// Uses `DashMap` for lock-free concurrent reads, matching the pattern
-/// used by `UcanContextState` in `runtime.rs`.
-static BRIDGE_STATE: OnceLock<DashMap<String, BridgeContextState>> = OnceLock::new();
-
-/// Returns a reference to the bridge state registry, initializing on first access.
-fn bridge_state_registry() -> &'static DashMap<String, BridgeContextState> {
-    BRIDGE_STATE.get_or_init(DashMap::new)
-}
-
-/// Removes per-context bridge state on context close, preventing unbounded
-/// memory growth in long-running processes. Called from `context_close`.
-pub(crate) fn remove_bridge_state(context_id: &str) {
-    bridge_state_registry().remove(context_id);
-}
+//
+// `BridgeContextState`, `bridge_state_registry()`, and `remove_bridge_state()`
+// are defined in `scp_ffi_common::bridge_state` and imported above.
 
 // ---------------------------------------------------------------------------
 // Result types
