+++ crates/scp-node/src/main.rs
@@ -30,9 +30,8 @@ use scp_platform::EncryptedStorage;
 use scp_platform::sqlite::{SqliteKeyCustody, SqliteStorage};
 use scp_platform::testing::{InMemoryKeyCustody, InMemoryStorage};
 use scp_platform::traits::Storage;
-use scp_transport::native::server::{RelayConfig, RelayServer};
-use scp_transport::native::storage::BlobStorageBackend;
-use tracing_subscriber::EnvFilter;
+use scp_transport::native::server::RelayServer;
+use scp_transport::startup;
 
 // ---------------------------------------------------------------------------
 // CLI argument parsing
@@ -135,17 +134,7 @@ ENVIRONMENT VARIABLES:
 // Environment variable helpers
 // ---------------------------------------------------------------------------
 
-/// Reads an environment variable and parses it, returning the default on
-/// absence or parse failure (with a warning).
-fn env_or<T: std::str::FromStr>(name: &str, default: T) -> T {
-    match env::var(name) {
-        Ok(val) => val.parse().unwrap_or_else(|_| {
-            tracing::warn!(var = name, value = %val, "invalid value, using default");
-            default
-        }),
-        Err(_) => default,
-    }
-}
+// `env_or` is provided by `scp_transport::startup::env_or`.
 
 // ---------------------------------------------------------------------------
 // Storage path resolution
@@ -326,46 +315,13 @@ impl<S: Storage + 'static> SequenceStore for StorageSequenceStore<S> {
 // Tracing
 // ---------------------------------------------------------------------------
 
-/// Initializes the `tracing` subscriber.
-fn init_tracing() {
-    let default_level = env::var("SCP_RELAY_LOG_LEVEL").unwrap_or_else(|_| "info".into());
-    let filter = EnvFilter::try_from_default_env().unwrap_or_else(|_| {
-        EnvFilter::try_new(&default_level).unwrap_or_else(|_| EnvFilter::new("info"))
-    });
-
-    let format = env::var("SCP_RELAY_LOG_FORMAT").unwrap_or_else(|_| "pretty".into());
-
-    if format == "json" {
-        tracing_subscriber::fmt()
-            .json()
-            .with_env_filter(filter)
-            .init();
-    } else {
-        tracing_subscriber::fmt().with_env_filter(filter).init();
-    }
-}
+// `init_tracing` is provided by `scp_transport::startup::init_tracing`.
 
 // ---------------------------------------------------------------------------
 // Relay config from env
 // ---------------------------------------------------------------------------
 
-/// Builds a [`RelayConfig`] from `SCP_RELAY_*` environment variables.
-fn relay_config_from_env() -> RelayConfig {
-    let bind_addr: SocketAddr = env_or(
-        "SCP_RELAY_BIND_ADDR",
-        SocketAddr::from(([0, 0, 0, 0], 9000)),
-    );
-
-    RelayConfig {
-        bind_addr,
-        max_blob_size: env_or("SCP_RELAY_MAX_BLOB_SIZE", 262_144),
-        max_blob_ttl: env_or("SCP_RELAY_MAX_BLOB_TTL", 604_800),
-        max_total_connections: env_or("SCP_RELAY_MAX_CONNECTIONS", 1_000),
-        max_connections_per_ip: env_or("SCP_RELAY_MAX_CONNECTIONS_PER_IP", 10),
-        rate_limit_publishes_per_second: env_or("SCP_RELAY_RATE_LIMIT", 100),
-        ..RelayConfig::default()
-    }
-}
+// `relay_config_from_env` is provided by `scp_transport::startup::relay_config_from_env`.
 
 // ---------------------------------------------------------------------------
 // Self-signed TLS provider (development mode)
@@ -395,130 +351,17 @@ impl TlsProvider for SelfSignedTlsProvider {
 }
 
 // ---------------------------------------------------------------------------
-// Health check
+// Health check + shutdown signal
 // ---------------------------------------------------------------------------
 
-/// Runs the `--health` probe: attempts a TCP connection to `addr` and
-/// exits with 0 on success, 1 on failure.
-async fn health_check(addr: SocketAddr) {
-    match tokio::net::TcpStream::connect(addr).await {
-        Ok(_) => std::process::exit(0),
-        Err(_) => std::process::exit(1),
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Shutdown signal
-// ---------------------------------------------------------------------------
-
-/// Waits for either SIGINT (`ctrl_c`) or SIGTERM.
-async fn shutdown_signal() {
-    let ctrl_c = tokio::signal::ctrl_c();
-
-    #[cfg(unix)]
-    {
-        let mut sigterm = tokio::signal::unix::signal(tokio::signal::unix::SignalKind::terminate())
-            .unwrap_or_else(|_| {
-                tokio::signal::unix::signal(tokio::signal::unix::SignalKind::interrupt())
-                    .unwrap_or_else(|_| std::process::exit(1))
-            });
-        tokio::select! {
-            _ = ctrl_c => {}
-            _ = sigterm.recv() => {}
-        }
-    }
-
-    #[cfg(not(unix))]
-    {
-        let _ = ctrl_c.await;
-    }
-}
+// `health_check` and `shutdown_signal` are provided by
+// `scp_transport::startup`.
 
 // ---------------------------------------------------------------------------
 // Relay blob storage from env
 // ---------------------------------------------------------------------------
 
-/// Valid backend names for error messages.
-const VALID_BLOB_BACKENDS: &str = "sqlite, redb, postgres, s3, memory";
-
-/// Constructs the blob storage backend from environment configuration.
-///
-/// Reads `SCP_RELAY_STORAGE_BACKEND` (default: `sqlite`) and delegates to the
-/// appropriate backend constructor. Exits on misconfiguration with a
-/// descriptive error naming the valid options.
-///
-/// This mirrors the logic in `scp-relay/src/main.rs::storage_from_env()`.
-async fn blob_storage_from_env() -> BlobStorageBackend {
-    let backend = env::var("SCP_RELAY_STORAGE_BACKEND")
-        .unwrap_or_else(|_| "sqlite".to_owned())
-        .to_lowercase();
-
-    match backend.as_str() {
-        "sqlite" => {
-            let path =
-                env::var("SCP_RELAY_STORAGE_PATH").unwrap_or_else(|_| "./scp-relay.db".to_owned());
-            let path = PathBuf::from(path);
-            tracing::info!(path = %path.display(), "using sqlite blob storage");
-            BlobStorageBackend::sqlite(&path).unwrap_or_else(|e| {
-                tracing::error!(error = %e, path = %path.display(), "failed to open sqlite blob storage");
-                std::process::exit(1);
-            })
-        }
-        "redb" => {
-            let path = env::var("SCP_RELAY_STORAGE_PATH")
-                .unwrap_or_else(|_| "./scp-relay.redb".to_owned());
-            let path = PathBuf::from(path);
-            tracing::info!(path = %path.display(), "using redb blob storage");
-            BlobStorageBackend::redb(&path).unwrap_or_else(|e| {
-                tracing::error!(error = %e, path = %path.display(), "failed to open redb blob storage");
-                std::process::exit(1);
-            })
-        }
-        "postgres" => {
-            let Ok(url) = env::var("SCP_RELAY_DATABASE_URL") else {
-                eprintln!(
-                    "error: SCP_RELAY_STORAGE_BACKEND=postgres requires SCP_RELAY_DATABASE_URL to be set"
-                );
-                std::process::exit(1);
-            };
-            tracing::info!("using postgres blob storage");
-            let store = scp_transport::native::postgres_blob::PostgresBlobStore::open(&url)
-                .await
-                .unwrap_or_else(|e| {
-                    tracing::error!(error = %e, "failed to connect to postgres");
-                    std::process::exit(1);
-                });
-            BlobStorageBackend::Postgres(store)
-        }
-        "s3" => {
-            let Ok(bucket) = env::var("SCP_RELAY_S3_BUCKET") else {
-                eprintln!(
-                    "error: SCP_RELAY_STORAGE_BACKEND=s3 requires SCP_RELAY_S3_BUCKET to be set"
-                );
-                std::process::exit(1);
-            };
-            let prefix = env::var("SCP_RELAY_S3_PREFIX").unwrap_or_else(|_| "blobs/".to_owned());
-            tracing::info!(bucket = %bucket, prefix = %prefix, "using s3 blob storage");
-            let store = scp_transport::native::s3_blob::S3BlobStore::open(&bucket, &prefix)
-                .await
-                .unwrap_or_else(|e| {
-                    tracing::error!(error = %e, "failed to initialize s3 storage");
-                    std::process::exit(1);
-                });
-            BlobStorageBackend::S3(store)
-        }
-        "memory" => {
-            tracing::warn!("using in-memory blob storage — all data will be lost on restart");
-            BlobStorageBackend::in_memory()
-        }
-        other => {
-            eprintln!(
-                "error: unknown storage backend '{other}'. Valid options: {VALID_BLOB_BACKENDS}"
-            );
-            std::process::exit(1);
-        }
-    }
-}
+// `storage_from_env` is provided by `scp_transport::startup::storage_from_env`.
 
 // ---------------------------------------------------------------------------
 // Relay-only mode
@@ -526,7 +369,7 @@ async fn blob_storage_from_env() -> BlobStorageBackend {
 
 /// Runs a bare relay server (same as `scp-relay` binary).
 async fn run_relay_only() {
-    let config = relay_config_from_env();
+    let config = startup::relay_config_from_env();
     tracing::info!(
         bind_addr = %config.bind_addr,
         max_blob_size = config.max_blob_size,
@@ -534,7 +377,7 @@ async fn run_relay_only() {
         "starting scp-node in relay-only mode"
     );
 
-    let storage = Arc::new(blob_storage_from_env().await);
+    let storage = Arc::new(startup::storage_from_env().await);
     let server = RelayServer::new(config, storage);
 
     let (handle, local_addr) = match server.start().await {
@@ -547,7 +390,7 @@ async fn run_relay_only() {
 
     tracing::info!(addr = %local_addr, "relay listening");
 
-    shutdown_signal().await;
+    startup::shutdown_signal().await;
 
     tracing::info!("shutdown signal received, stopping relay");
     handle.shutdown();
@@ -804,7 +647,7 @@ fn require_domain() -> String {
 
 /// Reads the HTTP bind address from env or returns the default.
 fn node_http_addr() -> SocketAddr {
-    env_or("SCP_NODE_BIND_ADDR", SocketAddr::from(([0, 0, 0, 0], 9000)))
+    startup::env_or("SCP_NODE_BIND_ADDR", SocketAddr::from(([0, 0, 0, 0], 9000)))
 }
 
 /// Builds a [`PkarrDhtClient`] from env configuration.
@@ -873,7 +716,7 @@ async fn run_node_with<
         .map(|v| v == "1" || v == "true")
         .unwrap_or(false);
 
-    let projection_rate: u32 = env_or(
+    let projection_rate: u32 = startup::env_or(
         "SCP_NODE_PROJECTION_RATE_LIMIT",
         scp_node::DEFAULT_PROJECTION_RATE_LIMIT,
     );
@@ -949,7 +792,7 @@ async fn run_node_with<
     // Install Prometheus metrics recorder and add /metrics endpoint (#1467).
     let metrics_router = install_metrics_recorder();
 
-    if let Err(e) = node.serve(metrics_router, shutdown_signal()).await {
+    if let Err(e) = node.serve(metrics_router, startup::shutdown_signal()).await {
         tracing::error!(error = %e, "application node exited with error");
         std::process::exit(1);
     }
@@ -1008,21 +851,21 @@ async fn main() {
     // --health: probe the appropriate bind address and exit.
     if config.health {
         let addr: SocketAddr = if config.relay_only {
-            env_or(
+            startup::env_or(
                 "SCP_RELAY_BIND_ADDR",
                 SocketAddr::from(([127, 0, 0, 1], 9000)),
             )
         } else {
-            env_or(
+            startup::env_or(
                 "SCP_NODE_BIND_ADDR",
                 SocketAddr::from(([127, 0, 0, 1], 9000)),
             )
         };
-        health_check(addr).await;
+        startup::health_check(addr).await;
         return;
     }
 
-    init_tracing();
+    startup::init_tracing();
 
     if config.relay_only {
         run_relay_only().await;
